+++ src/main/java/com/example/dreamvalutbackend/domain/genre/service/GenreService.java
@@ -13,9 +13,9 @@ import org.springframework.transaction.annotation.Transactional;
 import com.example.dreamvalutbackend.domain.genre.controller.response.GenreResponseDto;
 import com.example.dreamvalutbackend.domain.genre.controller.response.GenreWithTracksOverviewResponseDto;
 import com.example.dreamvalutbackend.domain.genre.controller.response.GenreWithTracksResponseDto;
-import com.example.dreamvalutbackend.domain.genre.controller.response.TrackOverviewResponseDto;
 import com.example.dreamvalutbackend.domain.genre.domain.Genre;
 import com.example.dreamvalutbackend.domain.genre.repository.GenreRepository;
+import com.example.dreamvalutbackend.domain.track.controller.response.TrackOverviewResponseDto;
 import com.example.dreamvalutbackend.domain.track.controller.response.TrackResponseDto;
 import com.example.dreamvalutbackend.domain.track.domain.Track;
 import com.example.dreamvalutbackend.domain.track.domain.TrackDetail;
